+++ cmd/image.go
@@ -7,7 +7,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
 
-	"github.com/openshift/cluster-version-operator/pkg/cvo"
+	"github.com/openshift/cluster-version-operator/pkg/payload"
 )
 
 var (
@@ -31,7 +31,7 @@ func runImageCmd(cmd *cobra.Command, args []string) {
 	if len(args) == 0 {
 		glog.Fatalf("missing command line argument short-name")
 	}
-	image, err := cvo.ImageForShortName(args[0])
+	image, err := payload.ImageForShortName(args[0])
 	if err != nil {
 		glog.Fatalf("error: %v", err)
 	}
