+++ pkg/cvo/updatepayload.go
@@ -1,173 +1,28 @@
 package cvo
 
 import (
-	"bytes"
 	"context"
 	"crypto/md5"
 	"encoding/base64"
 	"fmt"
-	"hash/fnv"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 
-	"k8s.io/client-go/kubernetes"
-
-	"github.com/golang/glog"
-	imagev1 "github.com/openshift/api/image/v1"
-	"github.com/pkg/errors"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	randutil "k8s.io/apimachinery/pkg/util/rand"
-	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/client-go/kubernetes"
 	"k8s.io/utils/pointer"
 
 	configv1 "github.com/openshift/api/config/v1"
-	"github.com/openshift/cluster-version-operator/lib"
 	"github.com/openshift/cluster-version-operator/lib/resourcebuilder"
-	"github.com/openshift/cluster-version-operator/lib/resourceread"
+	"github.com/openshift/cluster-version-operator/pkg/payload"
 )
-
-type updatePayload struct {
-	ReleaseImage   string
-	ReleaseVersion string
-	// XXX: cincinatti.json struct
-
-	ImageRef *imagev1.ImageStream
-
-	// manifestHash is a hash of the manifests included in this payload
-	ManifestHash string
-	Manifests    []lib.Manifest
-}
-
-const (
-	defaultUpdatePayloadDir = "/"
-	targetUpdatePayloadsDir = "/etc/cvo/updatepayloads"
-
-	cvoManifestDir     = "manifests"
-	releaseManifestDir = "release-manifests"
-
-	cincinnatiJSONFile  = "release-metadata"
-	imageReferencesFile = "image-references"
-)
-
-type payloadTasks struct {
-	idir       string
-	preprocess func([]byte) ([]byte, error)
-	skipFiles  sets.String
-}
-
-func loadUpdatePayloadMetadata(dir, releaseImage string) (*updatePayload, []payloadTasks, error) {
-	glog.V(4).Infof("Loading updatepayload from %q", dir)
-	if err := validateUpdatePayload(dir); err != nil {
-		return nil, nil, err
-	}
-	var (
-		cvoDir     = filepath.Join(dir, cvoManifestDir)
-		releaseDir = filepath.Join(dir, releaseManifestDir)
-	)
-
-	// XXX: load cincinnatiJSONFile
-	cjf := filepath.Join(releaseDir, cincinnatiJSONFile)
-	// XXX: load imageReferencesFile
-	irf := filepath.Join(releaseDir, imageReferencesFile)
-	imageRefData, err := ioutil.ReadFile(irf)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	imageRef, err := resourceread.ReadImageStreamV1(imageRefData)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "invalid image-references data %s", irf)
-	}
-
-	mrc := manifestRenderConfig{ReleaseImage: releaseImage}
-	tasks := []payloadTasks{{
-		idir:       cvoDir,
-		preprocess: func(ib []byte) ([]byte, error) { return renderManifest(mrc, ib) },
-		skipFiles:  sets.NewString(),
-	}, {
-		idir:       releaseDir,
-		preprocess: nil,
-		skipFiles:  sets.NewString(cjf, irf),
-	}}
-	return &updatePayload{ImageRef: imageRef, ReleaseImage: releaseImage, ReleaseVersion: imageRef.Name}, tasks, nil
-}
-
-func loadUpdatePayload(dir, releaseImage string) (*updatePayload, error) {
-	payload, tasks, err := loadUpdatePayloadMetadata(dir, releaseImage)
-	if err != nil {
-		return nil, err
-	}
-
-	var manifests []lib.Manifest
-	var errs []error
-	for _, task := range tasks {
-		files, err := ioutil.ReadDir(task.idir)
-		if err != nil {
-			return nil, err
-		}
-
-		for _, file := range files {
-			if file.IsDir() {
-				continue
-			}
-
-			switch filepath.Ext(file.Name()) {
-			case ".yaml", ".yml", ".json":
-			default:
-				continue
-			}
-
-			p := filepath.Join(task.idir, file.Name())
-			if task.skipFiles.Has(p) {
-				continue
-			}
-
-			raw, err := ioutil.ReadFile(p)
-			if err != nil {
-				errs = append(errs, errors.Wrapf(err, "error reading file %s", file.Name()))
-				continue
-			}
-			if task.preprocess != nil {
-				raw, err = task.preprocess(raw)
-				if err != nil {
-					errs = append(errs, errors.Wrapf(err, "error running preprocess on %s", file.Name()))
-					continue
-				}
-			}
-			ms, err := lib.ParseManifests(bytes.NewReader(raw))
-			if err != nil {
-				errs = append(errs, errors.Wrapf(err, "error parsing %s", file.Name()))
-				continue
-			}
-			manifests = append(manifests, ms...)
-		}
-	}
-
-	agg := utilerrors.NewAggregate(errs)
-	if agg != nil {
-		return nil, &updateError{
-			Reason:  "UpdatePayloadIntegrity",
-			Message: fmt.Sprintf("Error loading manifests from %s: %v", dir, agg.Error()),
-		}
-	}
-
-	hash := fnv.New64()
-	for _, manifest := range manifests {
-		hash.Write(manifest.Raw)
-	}
-
-	payload.ManifestHash = base64.URLEncoding.EncodeToString(hash.Sum(nil))
-	payload.Manifests = manifests
-	return payload, nil
-}
 
 func (optr *Operator) defaultPayloadDir() string {
 	if len(optr.payloadDir) == 0 {
-		return defaultUpdatePayloadDir
+		return payload.DefaultPayloadDir
 	}
 	return optr.payloadDir
 }
@@ -183,6 +38,10 @@ func (optr *Operator) defaultPayloadRetriever() PayloadRetriever {
 		workingDir:   targetUpdatePayloadsDir,
 	}
 }
+
+const (
+	targetUpdatePayloadsDir = "/etc/cvo/updatepayloads"
+)
 
 type payloadRetriever struct {
 	// releaseImage and payloadDir are the default payload identifiers - updates that point
@@ -209,7 +68,7 @@ func (r *payloadRetriever) RetrievePayload(ctx context.Context, update configv1.
 
 	tdir, err := r.targetUpdatePayloadDir(ctx, update)
 	if err != nil {
-		return "", &updateError{
+		return "", &payload.UpdateError{
 			Reason:  "UpdatePayloadRetrievalFailed",
 			Message: fmt.Sprintf("Unable to download and prepare the update: %v", err),
 		}
@@ -223,7 +82,7 @@ func (r *payloadRetriever) targetUpdatePayloadDir(ctx context.Context, update co
 	payloadHash := base64.RawURLEncoding.EncodeToString(hash.Sum(nil))
 
 	tdir := filepath.Join(r.workingDir, payloadHash)
-	err := validateUpdatePayload(tdir)
+	err := payload.ValidateDirectory(tdir)
 	if os.IsNotExist(err) {
 		// the dirs don't exist, try fetching the payload to tdir.
 		err = r.fetchUpdatePayloadToDir(ctx, tdir, update)
@@ -233,33 +92,10 @@ func (r *payloadRetriever) targetUpdatePayloadDir(ctx context.Context, update co
 	}
 
 	// now that payload has been loaded check validation.
-	if err := validateUpdatePayload(tdir); err != nil {
+	if err := payload.ValidateDirectory(tdir); err != nil {
 		return "", err
 	}
 	return tdir, nil
-}
-
-func validateUpdatePayload(dir string) error {
-	// XXX: validate that cincinnati.json is correct
-	// 		validate image-references files is correct.
-
-	// make sure cvo and release manifests dirs exist.
-	_, err := os.Stat(filepath.Join(dir, cvoManifestDir))
-	if err != nil {
-		return err
-	}
-	releaseDir := filepath.Join(dir, releaseManifestDir)
-	_, err = os.Stat(releaseDir)
-	if err != nil {
-		return err
-	}
-
-	// make sure image-references file exists in releaseDir
-	_, err = os.Stat(filepath.Join(releaseDir, imageReferencesFile))
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 func (r *payloadRetriever) fetchUpdatePayloadToDir(ctx context.Context, dir string, update configv1.Update) error {
@@ -332,12 +168,12 @@ func (r *payloadRetriever) fetchUpdatePayloadToDir(ctx context.Context, dir stri
 // `mkdir -p <target dir> && mv <default release manifest dir> <target release manifests dir>`
 func copyPayloadCmd(tdir string) string {
 	var (
-		fromCVOPath = filepath.Join(defaultUpdatePayloadDir, cvoManifestDir)
-		toCVOPath   = filepath.Join(tdir, cvoManifestDir)
+		fromCVOPath = filepath.Join(payload.DefaultPayloadDir, payload.CVOManifestDir)
+		toCVOPath   = filepath.Join(tdir, payload.CVOManifestDir)
 		cvoCmd      = fmt.Sprintf("mkdir -p %s && mv %s %s", tdir, fromCVOPath, toCVOPath)
 
-		fromReleasePath = filepath.Join(defaultUpdatePayloadDir, releaseManifestDir)
-		toReleasePath   = filepath.Join(tdir, releaseManifestDir)
+		fromReleasePath = filepath.Join(payload.DefaultPayloadDir, payload.ReleaseManifestDir)
+		toReleasePath   = filepath.Join(tdir, payload.ReleaseManifestDir)
 		releaseCmd      = fmt.Sprintf("mkdir -p %s && mv %s %s", tdir, fromReleasePath, toReleasePath)
 	)
 	return fmt.Sprintf("%s && %s", cvoCmd, releaseCmd)
